Mathematical Functions General Project
This program's primary purpose is to validate a bank card number (either credit or debit) by means of the Luhn Algorithm, which checks the positions from right to left (from last to first) of said character string containing the mentioned number, determining that if the relative position studied is odd, the number obtained is added to an accumulator, and if said position studied is even, the number obtained is multiplied by two and a unique number is obtained by adding both digits that resulted from said multiplication. The effective sum of the accumulated sums of the odd values ​​and the even values ​​with said nuance, must be divisible by ten for the number to be valid, otherwise it will be incorrect.

+++ CreditCardNumber.cpp
@@ -207,8 +207,8 @@ int int_getADigit(const int& int_digit_number)
 int int_sumOfDigits(const std::string &str_CardNumber)
 	{
 		/* Preliminary declarations of variables. */
-		int int_CardNumber_Length = str_CardNumber.length();
-		int int_CardNumber_Size = str_CardNumber.size();
+		int int_CardNumber_Length = int(str_CardNumber.length());
+		int int_CardNumber_Size = (int) str_CardNumber.size();
 
 		/* Initial declaration of work variables. */
 		int int_counting_items = V_ZERO, int_counting_pos_items = V_ZERO;
